Fixing things

+++ yaf_loader.h
@@ -87,7 +87,7 @@ extern zend_class_entry *yaf_loader_ce;
 int yaf_internal_autoload(char *file_name, uint name_len, char **directory);
 int yaf_loader_import(char *path, int len, int use_path);
 int yaf_register_autoloader(yaf_loader_t *loader);
-int yaf_loader_register_namespace_single(char *prefix, uint len);
+int yaf_loader_register_namespace_single(char *prefix, size_t len);
 yaf_loader_t *yaf_loader_instance(yaf_loader_t *this_ptr, zend_string *library_path, zend_string *global_path);
 
 extern PHPAPI int php_stream_open_for_zend_ex(const char *filename, zend_file_handle *handle, int mode);
